Update README; Avoid copy of existing dirs/files
Update README to reflect additions to config files.

Copy driver previously threw warnings about trying to copy already
existing directories/files. That's now avoid by skipping any that
already exist.

+++ src/main/scala/edu/umich/med/alzheimers/dicom/copy/Copier.scala
@@ -33,10 +33,13 @@ class Copier(
    */
   override def preVisitDirectory(dir: Path, attrs: BasicFileAttributes): FileVisitResult = {
     val target: Path = targetDirPath.resolve(sourceDirPath.relativize(dir))
+    val targetExists: Boolean = Files.exists(target)
 
     try {
-      Files.copy(dir, target, copyOptions: _*)
-      logger.info(s"Copy ${dir.toString} to ${target.toString}")
+      if (!targetExists) {
+        Files.copy(dir, target, copyOptions: _*)
+        logger.info(s"Copy ${dir.toString} to ${target.toString}")
+      }
     }
     catch {
       case e: DirectoryNotEmptyException =>
